requirements.txt
openai>=1.0.0
manim>=0.18.0
ffmpeg-python>=0.2.0
gtts>=2.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0
